refactor mapping tests into seperate service unit test
